fix bugs of using nevergrad

+++ config.py
@@ -18,7 +18,7 @@ def add_argument_group(name):
 glimpse_arg = add_argument_group('Glimpse Network Params')
 glimpse_arg.add_argument('--patch_size', type=int, default=8,
                          help='size of extracted patch at highest res')
-glimpse_arg.add_argument('--glimpse_scale', type=float, default=2,
+glimpse_arg.add_argument('--glimpse_scale', type=int, default=2,
                          help='scale of successive patches')
 glimpse_arg.add_argument('--num_patches', type=int, default=1,
                          help='# of downscaled patches per glimpse')
